requirements.txt
absl-py==2.1.0
altair==5.3.0
annotated-types==0.7.0
asgiref==3.8.1
astunparse==1.6.3
attrs==23.2.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
datetime-truncate==1.1.1
decorator==4.4.2
dj-database-url==2.2.0
Django==5.1
django-admin-charts==1.4.0
django-bower==5.2.0
django-jazzmin==3.0.0
django-light==0.1.0.post3
django-memoize==2.3.1
django-multiselectfield==0.1.13
django-nvd3==0.10.1
facenet-pytorch==2.6.0
fer==22.5.1
ffmpeg==1.4
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.2
google-pasta==0.2.0
googleapis-common-protos==1.65.0
googletrans==4.0.0rc1
grpcio==1.66.2
grpcio-status==1.66.2
gunicorn==23.0.0
h11==0.9.0
h2==3.2.0
h5py==3.11.0
hpack==3.0.0
hstspreload==2024.9.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
imageio==2.34.2
imageio-ffmpeg==0.5.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keras==3.4.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
ml-dtypes==0.4.0
moviepy==1.0.3
mpmath==1.3.0
namex==0.0.8
networkx==3.3
nltk==3.8.1
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.2
pillow==10.2.0
plotly==5.24.1
proglog==0.1.10
proto-plus==1.24.0
protobuf==5.28.2
psycopg2-binary==2.9.9
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-nvd3==0.16.0
python-slugify==8.0.4
python-stdnum==1.20
pytz==2024.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rfc3986==1.5.0
rich==13.7.1
rpds-py==0.19.1
rsa==4.9
setuptools==72.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sqlparse==0.5.1
streamlit==1.38.0
sympy==1.13.1
tenacity==8.5.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.4.0
text-unidecode==1.3
toml==0.10.2
toolz==0.12.1
torch==2.2.2
torchvision==0.17.2
tornado==6.4.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
waitress==3.0.0
watchdog==4.0.1
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0